requirements.txt
-i https://pypi.org/simple
appnope==0.1.0 ; sys_platform == 'darwin'
attrs==19.3.0
backcall==0.1.0
bleach==3.1.1
decorator==4.4.1
defusedxml==0.6.0
entrypoints==0.3
https://github.com/GillesPy2/GillesPy2/tarball/develop#egg=gillespy2
importlib-metadata==1.5.0 ; python_version < '3.8'
ipykernel==5.1.4
ipython-genutils==0.2.0
ipython==7.12.0
jedi==0.16.0
jinja2==2.11.1
json5==0.9.1
jsonschema==3.2.0
jupyter-client==6.0.0
jupyter-core==4.6.3
jupyterlab-server==1.0.6
jupyterlab==1.2.6
markupsafe==1.1.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.4
notebook==6.0.3
pandocfilters==1.4.2
parso==0.6.1
pexpect==4.8.0 ; sys_platform != 'win32'
pickleshare==0.7.5
prometheus-client==0.7.1
prompt-toolkit==3.0.3
ptyprocess==0.6.0 ; os_name != 'nt'
pygments==2.5.2
pyrsistent==0.15.7
python-dateutil==2.8.1
python-libsbml==5.18.0
pyzmq==19.0.0
send2trash==1.5.0
six==1.14.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.3
traitlets==4.3.3
wcwidth==0.1.8
webencodings==0.5.1
zipp==3.0.0
